Move components in App.js to their own individual files
