+++ src/App.js
@@ -1,46 +1,11 @@
 import React from 'react';
-import { Routes, Route, NavLink } from 'react-router-dom';
+import { Routes, Route } from 'react-router-dom';
+
 import './App.css';
+import NavLinks from './Navigation/NavLinks';
 import Firebase from './Pages/firebase';
-
-function List() {
-  return (
-    <>
-      <h1>List of Items</h1>
-    </>
-  );
-}
-
-function AddAnItem() {
-  return (
-    <>
-      <h1>Add an Item</h1>
-    </>
-  );
-}
-
-function NavLinks() {
-  return (
-    <nav style={{ position: 'fixed', bottom: 0 }}>
-      <NavLink
-        to="/"
-        style={({ isActive }) =>
-          isActive ? { fontWeight: 'bold' } : { fontWeight: 'normal' }
-        }
-      >
-        List of Items
-      </NavLink>
-      <NavLink
-        to="/add-an-item"
-        style={({ isActive }) =>
-          isActive ? { fontWeight: 'bold' } : { fontWeight: 'normal' }
-        }
-      >
-        Add an Item
-      </NavLink>
-    </nav>
-  );
-}
+import ItemList from './Pages/ItemList';
+import AddItem from './Pages/AddItem';
 
 function App() {
   return (
@@ -50,8 +15,8 @@ function App() {
       </header>
       <h1>Welcome to your Smart Shopping List!</h1>
       <Routes>
-        <Route path="/" element={<List />} />
-        <Route path="add-an-item" element={<AddAnItem />} />
+        <Route path="/item-list" element={<ItemList />} />
+        <Route path="/add-item" element={<AddItem />} />
       </Routes>
       <NavLinks />
     </div>
