+++ src/Pages/AddItem.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function AddItem() {
+  return (
+    <>
+      <h1>Add an Item</h1>
+    </>
+  );
+}
+
+export default AddItem;
