+++ src/Pages/ItemList.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function ItemList() {
+  return (
+    <>
+      <h1>List of Items</h1>
+    </>
+  );
+}
+
+export default ItemList;
